+++ apps/backoffice-v2/src/lib/blocks/components/DirectorsCallToAction/helpers.ts
@@ -2,7 +2,7 @@ import { AnyObject } from '@ballerine/ui';
 
 export const getRevisionReasonsForDocument = (
   { type, category }: AnyObject,
-  workflow: AnyObject,
+  revisionReasons: string[],
 ) => {
   if (category === 'proof_of_identity' && type === 'passport')
     return [
@@ -24,9 +24,5 @@ export const getRevisionReasonsForDocument = (
     ];
   }
 
-  return (
-    (workflow?.workflowDefinition?.contextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.revisionReason?.anyOf?.find(
-      ({ enum: enum_ }) => !!enum_,
-    )?.enum as string[]) || ([] as string[])
-  );
+  return revisionReasons;
 };